dont remove if env var is empty

+++ cleanup.sh
@@ -2,6 +2,6 @@
 
 set -euo pipefail
 
-if [[ -n ${TRUNK_TMPDIR+x} ]]; then
+if [[ -n ${TRUNK_TMPDIR+x} ]] && [[ -z ${TEST_GITHUB_EVENT_PATH} ]]; then
   rm -rf "${TRUNK_TMPDIR}"
 fi
